# skillmeat/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillMeat API",
    "description": "REST API for SkillMeat collection manager",
    "version": "0.1.0-alpha"
  },
  "paths": {
...
    "/api/v1/collections": {
      "get": {
        "tags": [
          "collections",
          "collections"
        ],
        "summary": "List all collections",
        "description": "Retrieve a paginated list of all collections with metadata",
        "operationId": "list_collections_api_v1_collections_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page (max 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of items per page (max 100)"
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor for pagination (next page)",
              "title": "After"
            },
            "description": "Cursor for pagination (next page)"
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved collections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/{collection_id}": {
      "get": {
        "tags": [
          "collections",
          "collections"
        ],
        "summary": "Get collection details",
        "description": "Retrieve detailed information about a specific collection",
        "operationId": "get_collection_api_v1_collections__collection_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/{collection_id}/artifacts": {
      "get": {
        "tags": [
          "collections",
          "collections"
        ],
        "summary": "List artifacts in collection",
        "description": "Retrieve a paginated list of artifacts within a specific collection",
        "operationId": "list_collection_artifacts_api_v1_collections__collection_id__artifacts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page (max 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of items per page (max 100)"
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor for pagination (next page)",
              "title": "After"
            },
            "description": "Cursor for pagination (next page)"
          },
          {
            "name": "artifact_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by artifact type (skill, command, agent)",
              "title": "Artifact Type"
            },
            "description": "Filter by artifact type (skill, command, agent)"
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionArtifactsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Artifact name",
            "examples": [
              "pdf-skill"
            ]
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Artifact type (skill, command, agent)",
            "examples": [
              "skill"
            ]
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "Current version",
            "examples": [
              "1.2.3"
            ]
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Source specification",
            "examples": [
              "anthropics/skills/pdf"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "source"
        ],
        "title": "ArtifactSummary",
        "description": "Summary of an artifact within a collection.\n\nLightweight artifact representation for collection listings."
      },
...
      "CollectionArtifactsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtifactSummary"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items for this page"
          },
          "page_info": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Pagination metadata"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page_info"
        ],
        "title": "CollectionArtifactsResponse",
        "description": "Paginated response for artifacts within a collection.\n\nReturns lightweight artifact summaries for efficient collection browsing."
      },
      "CollectionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CollectionResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items for this page"
          },
          "page_info": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Pagination metadata"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page_info"
        ],
        "title": "CollectionListResponse",
        "description": "Paginated response for collection listings.\n\nExtends the generic paginated response with collection-specific items."
      },
      "CollectionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Collection unique identifier",
            "examples": [
              "default"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Collection name",
            "examples": [
              "default"
            ]
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Collection format version",
            "examples": [
              "1.0.0"
            ]
          },
          "artifact_count": {
            "type": "integer",
            "title": "Artifact Count",
            "description": "Number of artifacts in collection",
            "examples": [
              5
            ]
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created",
            "description": "Collection creation timestamp"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "title": "Updated",
            "description": "Last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "artifact_count",
          "created",
          "updated"
        ],
        "title": "CollectionResponse",
        "description": "Response schema for a single collection.\n\nProvides complete collection metadata including artifact count\nand timestamps.",
        "example": {
          "artifact_count": 5,
          "created": "2024-11-16T12:00:00Z",
          "id": "default",
          "name": "default",
          "updated": "2024-11-16T15:30:00Z",
          "version": "1.0.0"
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type or code",
            "examples": [
              "NOT_FOUND"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message",
            "examples": [
              "Collection not found"
            ]
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional error details (optional)",
            "examples": [
              {
                "collection_id": "abc123"
              }
            ]
          },
          "validation_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Errors",
            "description": "Validation errors for invalid requests"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response envelope.\n\nAll API errors return this structure for consistent error handling\non the client side.",
        "example": {
          "error": "VALIDATION_ERROR",
          "message": "Invalid request parameters",
          "validation_errors": [
            {
              "field": "name",
              "message": "Field is required",
              "type": "value_error.missing"
            }
          ]
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageInfo": {
        "properties": {
          "has_next_page": {
            "type": "boolean",
            "title": "Has Next Page",
            "description": "Whether there are more items after this page",
            "examples": [
              true
            ]
          },
          "has_previous_page": {
            "type": "boolean",
            "title": "Has Previous Page",
            "description": "Whether there are items before this page",
            "examples": [
              false
            ]
          },
          "start_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Cursor",
            "description": "Cursor pointing to the first item in this page",
            "examples": [
              "Y3Vyc29yOjA="
            ]
          },
          "end_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Cursor",
            "description": "Cursor pointing to the last item in this page",
            "examples": [
              "Y3Vyc29yOjk="
            ]
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count",
            "description": "Total number of items (may be expensive to compute)",
            "examples": [
              42
            ]
          }
        },
        "type": "object",
        "required": [
          "has_next_page",
          "has_previous_page"
        ],
        "title": "PageInfo",
        "description": "Cursor-based pagination information.\n\nProvides information about the current page and navigation cursors\nfor efficient pagination of large datasets."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationErrorDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Name of the field that failed validation",
            "examples": [
              "name"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Error message describing the validation failure",
            "examples": [
              "Field is required"
            ]
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Type of validation error",
            "examples": [
              "value_error.missing"
            ]
          }
        },
        "type": "object",
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ValidationErrorDetail",
        "description": "Validation error detail for a single field.\n\nProvides structured information about validation errors\nthat occur during request processing."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}